+++ packages/core/src/runtime/Browser.ts
@@ -10,7 +10,8 @@ import {
 	Viewport,
 } from 'puppeteer'
 import DeviceDescriptors from 'puppeteer/DeviceDescriptors'
-import { Browser as BrowserInterface, EvaluateFn } from './types'
+import { EvaluateFn } from './types'
+import { Browser as BrowserInterface } from './IBrowser'
 import { NullableLocatable } from './Locatable'
 import CustomDeviceDescriptors from '../utils/CustomDeviceDescriptors'
 import { ElementHandle } from '../page/types'
