+++ packages/core/src/runtime/EvaluatedScript.ts
@@ -19,7 +19,8 @@ import {
 	RecoverWith,
 	extractStep,
 } from './Step'
-import { SuiteDefinition, Browser } from './types'
+import { SuiteDefinition } from './types'
+import { Browser } from './IBrowser'
 import Test from './Test'
 import { mustCompileFile } from '../TestScript'
 import { TestScriptError, TestScriptErrorMapper } from '../TestScriptError'
